[STEP] Sylvain save in data base

+++ src/main/java/fr/nwwdjavaspringboot/model/NWDBusiness/NWDBasicModel.java
@@ -1,8 +1,10 @@
 package fr.nwwdjavaspringboot.model.NWDBusiness;
 
+import java.math.BigInteger;
+
 public class NWDBasicModel extends NWDDatabaseBasicModel{
 
-    public long projectId;
+    public BigInteger projectId;
 
     public void CopyFrom(NWDBasicModel sOther){
         super.CopyFrom(sOther);
