+++ src/main/java/fr/nwwdjavaspringboot/model/NWDBusiness/exchanges/basic/NWDBasicExchange.java
@@ -6,13 +6,14 @@ import fr.nwwdjavaspringboot.model.NWDBusiness.NWDEnvironmentKind;
 import lombok.NoArgsConstructor;
 
 import java.io.Serializable;
+import java.math.BigInteger;
 
 @JsonIgnoreProperties(ignoreUnknown = true)
 @NoArgsConstructor
 public abstract class NWDBasicExchange implements Serializable {
 //    public String Dll = NWDVersionDll.Version;
 //    public String IdName  = NWDRandom.RandomString(8);
-    public long projectId;
+    public BigInteger projectId;
     public NWDEnvironmentKind environment;
     public String stamp;
     public String hash;
