+++ src/main/java/fr/nwwdjavaspringboot/model/NWDBusiness/facade/INWDProjectInformation.java
@@ -2,9 +2,11 @@ package fr.nwwdjavaspringboot.model.NWDBusiness.facade;
 
 import fr.nwwdjavaspringboot.model.NWDBusiness.NWDEnvironmentKind;
 
+import java.math.BigInteger;
+
 public interface INWDProjectInformation {
 
     public String GetProjectInformationInstanceName();
-    public long GetProjectId();
+    public BigInteger GetProjectId();
     public NWDEnvironmentKind GetProjectEnvironment();
 }
