+++ src/main/java/fr/nwwdjavaspringboot/model/NWDBusiness/facade/INWDProjectKey.java
@@ -2,7 +2,9 @@ package fr.nwwdjavaspringboot.model.NWDBusiness.facade;
 
 import fr.nwwdjavaspringboot.model.NWDBusiness.NWDEnvironmentKind;
 
+import java.math.BigInteger;
+
 public interface INWDProjectKey {
     public String getProjectKeyInstanceName();
-    public String getProjectKey(long sProjectId, NWDEnvironmentKind sEnvironmentKind);
+    public String getProjectKey(BigInteger sProjectId, NWDEnvironmentKind sEnvironmentKind);
 }
